backend/requirements.txt
Flask==2.1.0
flask_cors==3.0.10
flask_login==0.6.2
Werkzeug==2.1.0
SQLAlchemy==1.4.46
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
flask_jwt_extended>=4.4.0
flask-sqlalchemy
